requirements.txt
flask
flask-cors
openai
python-dotenv
sqlalchemy
psycopg2-binary
flask-jwt-extended
requests
gunicorn